Create FirebaseMessagingDelegate class
